Added copying of environment-specific resources to target installation.
In the event one would like to have extra resources copied to wherever in
the resulting file there is a folder named "deployment" having the
following structure:

deployment
  + common
  + development
  + test
  + production

The common subfolder contains all the files that should be copied
regardles of the environment.
The rest of the folders should contain files for specific environments.
For example if you're building the production deployment bundle you'd
execute

...> grails prod bundle

and the actual environment will be "production". It works even with custom
environments - you just need to name the folder properly.